Created the top navigation
Created the initial top nvaigation. Everything is aligned to the left and is floating from the left. still unsure how to actually center the nav - attempted to create two different UL (classes indicating one is housed on the left and the other is housed on the right). Did not work so we will see what ends up happening. Styled to be consistent. Can I try and create IDs or classes for specific elements inside of the nav list?