Wrapping up on flavor page
Built out the flavor column cards to include images and text. Also modified the h2 and secondary class of paragraph text, as well as modifying  a button (tertiaryButton) to be akin to my initial  design. Had to create a "spacer" div to stop the cards from being affected by the button (i think because it was floating). I am still really unsure how to use vectors - and how to properly show mouseover effects only with css.