Flushed out pages and added more
Finished filling in the flavors page with other pieces of content (the section navigation, hover effects on the cards, links, etc.) and replicated the changes across the board for the Glaze and Quantity page. Still have not touched vectors which I assume will be difficult, and have no idea how I want to do the overlay yet. That will be interesting. I feel like I am having to "dumb things down" because my only constraints are myself and the deadline.