Empty cart and multiple fixes
To maintain consistency with other buttons, made the top navigation and tertiary buttons consistent with the same typeface as the primary buttons (lato).

Also created and added the empty cart, and manipulated the current classes of content to stylize the cart page as close as possible to prototype. Unable to create an actual overlay, so created a split content page that houses a static image of the quanitity page on the left side of the page. Tempted to build out specific empty cart pages to redirect each "close" button back to their specific page until i implement JS to send users straight back to whatever page they were on. But for now, we roll with it. On to the filled cart and checkout form which is going to be SOOOOOO fun.