Created a new structure for the item creator
Played with the sturcture of the item creation so that the user stays on the same page at all times. I still cannot figure out the local storage which is stopping me from spanning the process across multiple pages like I intended. I have had to pivot my design to actually get make it usable, even though it is not what I was intending. It seems to me like this is more of a prototype then an interactive site, so I would like to figure out how to properly implement the different techniques needed to make it interactive rather than just a flushed out prototype.