Flushed some stuff out
Attempted to mimic the real world scenario I was going for. Built out the checkout flows within the flow-test page so that the data would not be lost. Also added the option to remove the created item, but it just hides it. I still have not been able to add an element or properly remove an element, and i believe I would need to use local storage for that. I wonder if there is a way to save a copy of an element and let that sit inside of the div? That way any new changes would not be directly effecting a previous item that is created..